+++ test/fixtures/dot_shorthands/chain_index.input.dart
@@ -1,0 +1,3 @@
+enum Color { red, green, blue }
+
+Color colorAt(int i) => Color.values[i];