+++ test/fixtures/dot_shorthands/chain_property_access.input.dart
@@ -1,0 +1,3 @@
+enum Color { red, green, blue }
+
+Color firstColor() => Color.values.first;